--- a/devpod-automation/config.nu
+++ b/devpod-automation/config.nu
@@ -1,0 +1,5 @@
+let config = {
+    templates: "devpod-automation/templates"
+    scripts: "devpod-automation/scripts"
+    output_dir: "output"
+}