… .claude/commands/_base_prp_command.md
… .claude/commands/execute-prp.md
… .claude/commands/generate-prp-v2.py
… .claude/commands/generate-prp.md
… CLAUDE.md
… common.json
… context-engineering/docs/composite-template-system.md
… context-engineering/docs/environment-adapter-implementation.md
… context-engineering/docs/prp-builder-implementation.md
… context-engineering/docs/security-implementation.md
… context-engineering/docs/version-control-and-scalability-implementation.md
… context-engineering/execution-reports/README.md
… context-engineering/templates/config/python_config.yaml
… context-engineering/templates/config/rust_config.yaml
… context-engineering/templates/config/typescript_config.yaml
… context-engineering/templates/fragments/base_header.md
… context-engineering/templates/fragments/environment_setup.md
… context-engineering/templates/fragments/task_structure.md
… context-engineering/templates/fragments/validation_gates.md
… context-engineering/test_composite_template.py
… context-engineering/test_environment_adapter.py
… context-engineering/test_integrated_system.py
… context-engineering/test_prp_builder.py
… context-engineering/test_secure_command_system.py
… devpod-automation/provision.sh
… devpod-automation/scripts/devpod-generate.nu
… devpod-automation/templates/go/devcontainer.json
… devpod-automation/templates/python/devcontainer.json
… devpod-automation/templates/rust/devcontainer.json
… devpod-automation/templates/typescript/devcontainer.json